requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.2
pydantic-settings>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1
python-dotenv>=1.0.0
aerich>=0.7.2
tortoise-orm>=0.20.0
loguru>=0.7.2
setuptools>=75.8.0
aiomysql>=0.2.0
redis[hiredis]>=5.0.1
python-jose[cryptography]>=3.3.0
motor==3.3.2
aiohttp>=3.11.12 